refactor controllers

+++ src/main/java/com/travel_planner_be/travel/service/RouteService.java
@@ -1,24 +1,29 @@
 package com.travel_planner_be.travel.service;
 
+import com.travel_planner_be.travel.entity.Place;
 import com.travel_planner_be.travel.entity.Route;
+import com.travel_planner_be.travel.entity.User;
 import com.travel_planner_be.travel.repository.RouteRepository;
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
 @Service
+@RequiredArgsConstructor
 public class RouteService {
 
-    @Autowired
-    RouteRepository routeRepository;
+    private final RouteRepository routeRepository;
+    private final UserService userService;
+    private final PlaceService placeService;
 
-    public Route saveRoute(Route route) {
-        route.setId(UUID.randomUUID().toString());
 
-        return routeRepository.save(route);
-    }
     public List<Route> getRouteByUserId(String userId) {
         return routeRepository.findAllByUserId(userId);
     }
@@ -28,15 +33,51 @@ public class RouteService {
                 .orElse(null);
     }
 
-    public boolean cancelRoute(String routeId) {
+    public ResponseEntity<String> cancelRoute(String routeId) {
+        boolean isDeleted;
+        if (routeId == null || routeId.isEmpty()) {
+            return new ResponseEntity<>("Invalid routeId", HttpStatus.BAD_REQUEST);
+        }
         if (routeRepository.existsById(routeId)) {
             routeRepository.deleteById(routeId);
-            return true;
+            isDeleted = true;
         } else {
-            return false;
+            isDeleted = false;
+        }
+        if (isDeleted) {
+            return new ResponseEntity<>("Tour successfully canceled", HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>("Tour cancellation failed", HttpStatus.NOT_FOUND);
         }
     }
 
+     public ResponseEntity<Route> saveRoute(Route route){
+         route.setId(UUID.randomUUID().toString());
+         Route savedRoute = routeRepository.save(route);
+
+         Optional<User> optionalUser = userService.getUserById(route.getUserId());
+
+         for (String placeId : savedRoute.getPlaces()) {
+             Place selectedPlace = placeService.getPlace(placeId);
+             selectedPlace.setPopularityRate(selectedPlace.getPopularityRate() + 1);
+             placeService.updatePlace(selectedPlace);
+         }
+
+         if (optionalUser.isPresent()) {
+             User user = optionalUser.get();
+             if (user.getRoutes() == null) {
+                 user.setRoutes(new ArrayList<>());
+             }
+             user.getRoutes().add(savedRoute.getId());
+
+             userService.saveUser(user);
+
+             return new ResponseEntity<>(savedRoute, HttpStatus.OK);
+         } else {
+
+             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+         }
+     }
 
 
 }
